Home banner updates
Added several styles to test in banner.yaml.
Added option to flip the layout on mobile.
Added variables to the bitcoin logo SVG.
Changed yaml variable names to include the #. Otherwise only color variables that start with a letter would work and words like "transparent" also wouldn't.
Logo is now only shown if a fill or outline are provided.