Home banner tweaks
Added theme support for the SVG and mobile layout flip.
Made bitcoin logo clickable to change the theme.
Instead of a random theme, the the theme array is initially randomized and then clicking tthe logo linearly loops through the list to avoid seeing the same one over and over.
Fixed initial load (HTML is not available when JS runs for the first time, we may need to move the JS into the footer).
Added CSS easing for smoother theme switching.