First draft of a page with content guidelines
As we make more content public, we will see many diverse writing styles and perspectives. That is great. There are still a few reasons to have some guidelines, so readers can have a smooth experience and don't feel big disconnects whenever they click through to a different page. Here I tried to gather some thoughts on writing that we can iterate on.