Guide intro page & cleanup
Added a guide intro page with some first-draft copy. It contains placeholder links to the upcoming chapters that need to be updated as those are being worked on.
Removed various placeholder markdown pages.
Minor updates to CSS and a new picture.html include for better control over guide content layouts.