Moved content guidelines and style pages into resources folder
Renamed "Style" page to "Formatting", as style could also be interpreted as writing style.
Moved "Formatting" and "Content guidelines" into a "Resources" sub-nav of the guide.
Created a "Resources" intro page that summarizes sub-pages.
Minor clean-up of the "Formatting" page.
Removed a placeholder "wallet features" page that was floating around in the images folder.