Self-custody typo
Based on Daniel's suggestion of changing all cases on non-custody to self-custody I went over the guide to make this change. This was the only instance found on the current master branch. For those writing copy currently for the guide it would be good to make this change on current work before pushing PRs.