typo
Changed extended private key and public key entries to lowercase. This is the proper way to write these. Also removed the to do that was listed. We should open an issue if we want to add that not insert it here. Especially if we are planning a soft launch soon.