Updates the site logo to the seal
Changes the favicon and logos used in the mobile header, home page banner, and site footer.

The generic outline Bitcoin symbol was replaced with the community seal. I've run this design change past the community on Slack and several people individually. So far everyone seems to agree that this is a good direction, now let's see how we like it in the site itself.