Always display sub-sub pages in the sidebar nav
Removes the small arrow to expand sub-folders to reduce amount of clicking to be done. Instead, sub-sub pages are always visible. The hierarchy is still clear to viewers through the visual indentation.

Also minor tweaks to the visuals:
- Slightly tighter spacing on desktop
- Very light grey tone used now for sub-pages
- Added orange hover state for nav links

This is part of the changed discussed in #551. As it's pretty self-contained, I thought this could be a quick one to get out of the way.