Added note around localization considerations for recovery phrases
If you're writing Japanese or other languages that are very different from English, then the words in a recovery phrase will look pretty foreign to you. Muun has gotten this feedback and it was a reason why they were more comfortable using output descriptors.

Not sure if the glossary is the right place for this, but I think it would be good to add this somewhere. Also wonder if we have any recommendations on what to do instead?