Quick and dirty Lottie test for the home page banner
Just testing how complex it would be to implement Lottie for animations in the home page banner. This test breaks browsing other banners, just meant to be an exploration.

Overall, I think it works well. If we want to do this, we need to do a proper implementation that only loads the Lottie library when required, fix the banner theme navigation, etc.