Fixed Lottie implementation & used Jessica's banner animation
Lottie is now dynamically loaded when the currently active banner needs it.

Deleted the temporary astronaut animation I had uses for testing.

Implemented the Lottie animation for Jessica's banner. This needs fixing though. The static image had a white border that is missing in the animation, and some animation elements are images and should be vectors (for proper retina scaling and best file size).