Add notes about community association
Added a "usage note" to the community seal link/download on the "Design files" page. The note asks the reader to get consensus before associating the seal with a project more formally.

Second addition is a note on the projects page about adding projects/collaborations that points out that lines are a bit blurry and it's best to ask and get consensus.