Add content around address formatting and validation
Adds two paragraphs to the Address page in the Glossary.
- Visual formatting of addresses
- Validation of addresses

I could not find any good info on exactly how well address can be validated. In a recent podcast, someone said that in some formats, it is possible to identify which character is wrong. Maybe someone with more technical knowledge in this area could chime in.

Closes #107, closes #806