Adds an include for video links with small thumbnails
This allows us to more subtle promote video material in our content. We already have a big video link that covers the whole page width. Sometimes, we want to show videos in a more subtle way. This component set allows for showing single videos, or grids of videos. The layout adjusts based on screen size.