Adding various YouTube video references
We have really useful videos on lots of topics. With this PR, I am adding links with thumbnails to several pages where our videos can complement the content we have.

This is currently only looking at videos the community has created. We could expand this further to videos from others.