Adds test videos to the accessibility page
Mo and I did several accessibility tests on wallets during our research for the page. We summarized our findings in videos. These can be super helpful for interested designers to get an idea of what testing looks like, and what problems are typically found. So let's add the video links here to the page directly.

This came out of a discussion in Mondays jam session.