Adds required files for a Nostr profile
This is part of #470, setting up a Nostr profile for the community.

The `.well-known/nostr.json` allows us to use the handle `@bitcoin.design` (or `community@bitcoin.design`) in Nostr clients. It links the domain, name, and public key together, following the NIP-05 standard.
https://github.com/nostr-protocol/nips/blob/master/05.md

The added profile picture will be used for creating the profile event and posting it to relays, following NIP-01.
https://github.com/nostr-protocol/nips/blob/master/01.md#basic-event-kinds

Speaking of relays, I don't know which ones are "good", so I picked ones that appear like they would be well-visited. But let's get some community input. Which relays do you check or post to?