Add Nostr profile to the social media page
- Adds Nostr-specific copy to the social media page
- Minor pluralization copy changes to make the page about multiple accounts

Also fixing an issue with the nostr.json file. Convention appears to be to use an underscore in the file if you want the domain as a whole to be linked to a specific account. See [here](https://github.com/nostr-protocol/nips/blob/master/05.md#showing-just-the-domain-as-an-identifier). Some clients could not load the profile via @bitcoin.design, I hope this tweak fixes that.

I am using the nostr: URL scheme, although pretty much no one will have a client installed that picks this up. I added a mention of this right after the first time this scheme is used.