Add a link to the survey in the contribute section (#1017)
We did a survey on how people use the design guide, and Mo put together a great presentation on it. I'm adding a link in the contribute section for reference, right where we give people tips on finding ways to improve the guide. Then we can close the GitHub issue for the survey, as discussed in the comments.

See: https://github.com/BitcoinDesign/Meta/issues/320