Extract send fees into a page and add scenarios
Picking up work that Steyn did around various fee scenarios, and tried to marry it with what we have in the guide already. The "Send" page had a lot of fee-related info, but adding more there would have resulted in a very long page. That's why I extracting the fee content into a dedicated page that lives underneath the "Send" page.

Closes #997