Implements prefers-contrast media query (#1015)
* Implements prefers-contrast media query

This allows users to enable increased contrast in the accessibility settings of their operating system, and see that reflected in the website.

This is also a proof-of-concept for implementing this using CSS variables. It requires replacing $ variables (like $primary-color) with CSS variables (var(--primary-color). Benefit is that the CSS variables can dynamically change at runtime (not just during build time). This more dynamic system allows for the prefers contrast color change to only require 8 lines of code.

But, this needs some testing to ensure it all works cross-browser.

* Adjust the high-contrast color based on feedback