Updated gemfile
So GitHub auto-created PR #1072 as a security update. My local dev environment no longer worked with this. After around 6 hours of trying to fix it myself (always seemed to come down to Ruby compilation problems related to OpenSSL versions), I bought Ruby on Mac and used it to completely reset my local dev environment and move to a newer version of Ruby (3.2.3). Now it works again. We have not updated this lock file in years, maybe it's a good time to push this update so we use the latest version of Ruby and various dependencies?

But first we should have a few people test this. Maybe it's just me having all these problems.