Updated address character length (#1075)
This was brought up on Discord by zoop. I believe we got our initial information for this incorrect length range from the Bitcoin Wiki, which still reads 26-35. However, Taproot addresses can be 62 characters long. That page does not discuss Taproot, which is probably the reason for this difference.

Specific changes:

- "26-35" -> "26-62" for address length range
- Removed a link to the Wiki page about address formats, since it does not include Taproot
- Removed a link to the Bech32 Wiki page, because that link is next to "Taproot address" header, and that linked page does not discuss Taproot.

Just a few minor tweaks for accuracy here. Let me know if you see anything else that could be improved.