Contribute page revision and renaming (#1079)
* Contribute page revision and renaming

This is a larger revision of the contribute page. It is being renamed to "Join", which is more appropriate. It takes the pressure off from dedicating yourself ot helping out and opens to door to just starting by poking around. The page is now broken down into 7 milestones that slowly easy a newcomer into the community, open-source and bitcoin. It can be tricky to find your way around this ecosystem. Hopefully this journey-like approach and more casual tone makes it a little easier.

Various images from the projects page were also in the contribute page folder. I moved those to the projects page folder.

I also updated any links to "/contribute/" to the new destination at "/join/". A redirect for the old link is in place.

Closes #1053

* Apply suggestions from code review

Co-authored-by: Mogashni  <90313394+mouxdesign@users.noreply.github.com>

* Formatting fixes and two extra links in tune-in section

* Update guide/getting-started/why-design-for-bitcoin.md

Co-authored-by: Mogashni  <90313394+mouxdesign@users.noreply.github.com>

---------

Co-authored-by: Mogashni <90313394+mouxdesign@users.noreply.github.com>