Adds new page about human readable addresses
This is something we've discussed for a few weeks in the UX channel in Discord and turns the findings into a page.

I think there's more work to do in cross-linking this from all the right places. I updated a few spots, but there's probably more.

The UI mock-ups are more directional than implementation-ready, which I think is fine. But we could flesh this out further in a revision at a later point.