Usability and footer tweaks based on feedback
This addresses the super small easy-to-implement site feedback we got from Sven in the Open Source Design forum (https://discourse.opensourcedesign.net/t/new-contribute-join-page-in-the-bitcoin-design-community/3769/2).

- Adding Nostr and BitcoinTV logos to the footer
- Add a hover tip to the footer logo and the header search icon
- Active and visited states for footer links

Since Nostr is not a single service, the link goes to Nosta, which is a service meant for showing Nostr profiles. I am also the creator of Nosta, so am I self-serving by doing this? Let me know what you think, we can change the link elsewhere.