Adds the Coinbits personas to the user personas page (#1095)
* Adds the coinbits personas to the user personas page

Images of the personas are included to preserve the personas in case the original material disappears at some point.

Also adds a table of contents. The new personas are a bit hidden at the bottom. The table of contents makes it easier to see what's on the page.

* Tweak to force re-deploy

Failed because a download couldn't finish. But the download works for me. So trying again here.

* Updated link