Add Umami tracking script (#1129)
* Add Umami tracking script

As discussed in issue #1125, this adds the tracking script to our self-hosted instance of Umami. It is configured to only track on the bitcoin.design domain.

* Add a note about tracking on the join page

Could not find a better page to place this on. My logic is that it's helpful info for someone who intends to join the community.