Bubble Wallet mock-up revision and note
The naming conventions in the mock-ups seem to be throwing people off (which is not a first in the history of all the discussions we have had around the guide). This commit adds a special note above the first visual mock-up to explain the naming logic, and how other wallets can make different decisions. Mock-ups are tweaked accordingly.