Adds new page about human readable addresses (#1082)
* Adds new page about human readable addresses

This is something we've discussed for a few weeks in the UX channel in Discord and turns the findings into a page.

I think there's more work to do in cross-linking this from all the right places. I updated a few spots, but there's probably more.

The UI mock-ups are more directional than implementation-ready, which I think is fine. But we could flesh this out further in a revision at a later point.

* Updated domains in images

* Update guide/how-it-works/human-readable-addresses.md

Co-authored-by: Johns Beharry <johns@peakshift.com>

* Update guide/how-it-works/human-readable-addresses.md

Co-authored-by: Johns Beharry <johns@peakshift.com>

* Update guide/how-it-works/human-readable-addresses.md

Co-authored-by: Johns Beharry <johns@peakshift.com>

* Tweaks based on feedback

- Added note about the DNS proposal still being in discussion
- Updated general how it works image to use [prefix] instead of ₿
- Added note about PayNym directory not being open-source, and removed note about not using it due to centralization (should be clear from reading the text whether to use it or not)

* Apply suggestions from code review

Co-authored-by: Stephen DeLorme <stephen@d.elor.me>

* Address review feedback & minor updates

- The proposal has been assigned a BIP number in the mean time, so I updated the URL
- Clarified the use of the term custody as it relates to addresses
- Removed an incorrect sentence about caching
- Clarified that the proposal supports two formats, due to the optional ₿ prefix
- Added note how single-use addresses can be used, as long as there's a rotation mechanism in place
- Simplified note about responsibility to be more general, rather than mentioning legal and financial aspects. Lightened up that sentence generally
- Added note about informing user of changed payment info at time of initiating a payment
- Minor copy tweak in lightning address intro paragraph

* General page revision

This update should address a lot of the feedback that has come up in the past months.

- Revision of the address setup flow (which I thought I'd do in #1083, but here it is)
- Added notes about user-facing naming (probs will end up in discussion, as naming tends to do)
- Lots of small copy tweaks and additions
- Added a table of contents at the top
- Revision of diagrams for clarity
- Change b12 to lno to match the updated convention
- Updated backlink from the private key management intro page

Would be amazing to get some fresh reviews on this. There may also have been other changes to the proposal or learnings from real-world implementations that should be considered.

* Compressing images

Saving 4.2 of 6.8 MB.

* Apply suggestions from code review

Co-authored-by: Max <114083781+yashrajd@users.noreply.github.com>

* Update mapping diagram

Replace regular bitcoin address with silent payments address, per Yashraj's feedback.

* Bubble Wallet mock-up revision and note

The naming conventions in the mock-ups seem to be throwing people off (which is not a first in the history of all the discussions we have had around the guide). This commit adds a special note above the first visual mock-up to explain the naming logic, and how other wallets can make different decisions. Mock-ups are tweaked accordingly.

---------

Co-authored-by: Johns Beharry <johns@peakshift.com>
Co-authored-by: Stephen DeLorme <stephen@d.elor.me>
Co-authored-by: Max <114083781+yashrajd@users.noreply.github.com>