Lots of link fixes
Did a quick "rake test" as a follow-up on the last merge, just to see if there were any issues. Turns out there were over 300. So I went through and fixes broken links and file references.

Some of those must have been in there for a long time. Which is weird, because I do run the script here and there.