Copy iteration of the Getting Started chapter landing page (#1157)
This builds on the work done by @Brendaoduor in her Google Doc. I tweaked it a little bit further to make the text more about what the reader takeaways are in general, rather than summarizing the page content directly. I also tried to keep it very approachable. Curious what others think of this.

Closes #779