Add Gemini State of Crypto research report
It contains some interesting data across 6 countries regarding the ownership, attitudes, etc of crypto, with several points specifically on bitcoin.

This PR also moves the "Journal articles" section of the page up. Reason is that this section is getting content added frequently. The section previously at the top, "Journal articles" shows the latest addition from 2021. So when you start browsing the page from the top, you see 2021, and think the whole page is outdated right off the bat. I am wondering if we should restructure the page and get rid of the categorization altogether, and make it reverse chronological.